등록수저 및지 컴포넌트 나누기
