+++ package.json
@@ -28,7 +28,6 @@
     "firebase": "^9.6.10",
     "framer-motion": "^1.7.0",
     "graphql": "^16.3.0",
-    "jquery": "^3.6.0",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-bootstrap": "^2.4.0",
