+++ src/commons/imageBoard/upload.queries.ts
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const UPLOAD_BOARD_IMAGE = gql`
+  mutation uploadBoardImage($files: [Upload!]!) {
+    uploadBoardImage(files: $files)
+  }
+`;
