+++ src/commons/imageBoard/upload.styles.ts
@@ -1,0 +1,28 @@
+import styled from "@emotion/styled";
+
+export const UploadImage = styled.img`
+  object-fit: cover;
+  width: 40%;
+  height: 190px;
+  outline: none;
+  cursor: pointer;
+  border-radius: 10px;
+`;
+export const UploadButton = styled.div`
+  width: 35%;
+  height: 180px;
+  outline: none;
+  cursor: pointer;
+  display: flex;
+  justify-content: center;
+  padding: 9%;
+  margin: 2%;
+  border-radius: 10px;
+`;
+export const UploadIcon = styled.img`
+  width: 50%;
+  height: 90%;
+`;
+export const UploadFileHidden = styled.input`
+  display: none;
+`;
