+++ src/commons/imageBoard/upload.types.ts
@@ -1,0 +1,8 @@
+import { IQuery } from "../types/generated/types";
+
+export interface IImageUploadPageProps {
+  onChangeFileUrls: (fileUrl: string, index: number) => void;
+  index: number;
+  fileUrl: string;
+  data?: any;
+}
