+++ src/component/unit/boards/write/write.presenter.tsx
@@ -1,13 +1,9 @@
-// haeri 작업시작 22.05.12
 import * as S from "./write.styles";
 import { IWritePresenter } from "./write.types";
-import { Modal, Space } from "antd";
-import { formats, modules } from "../../../../commons/quill";
-import KeyWord from "../../../../commons/kakao/keyword/kakaomap.container";
 import Script from "next/script";
-import ImageBoardUpload from "../../../../commons/imageBoard";
-import { v4 as uuidv4 } from "uuid";
-import moment from "moment";
+import First from "./write.firstpresenter";
+import Second from "./write.secondpresenter";
+import Third from "./write.thirdpresenter";
 import "moment/locale/zh-cn";
 
 export default function WritePresenter(props: IWritePresenter) {
@@ -18,9 +14,7 @@ export default function WritePresenter(props: IWritePresenter) {
         src="//dapi.kakao.com/v2/maps/sdk.js?appkey=10933d05118bfc99d732e83a2814b76a&libraries=services&autoload=false"
         strategy="beforeInteractive"
       />
-
       <S.TopWrapper>
-        {/* STEP1 */}
         {props.activeStep === "first" && (
           <S.TopOn>
             <S.TopRow>
@@ -32,12 +26,10 @@ export default function WritePresenter(props: IWritePresenter) {
         )}
         {props.activeStep !== "first" && (
           <S.TopOff>
-            <S.StepName>STEP 1</S.StepName>
+            <S.StepName onClick={props.onClickSecondPrev}>STEP 1</S.StepName>
             <S.StepDetail> 거주 기간, 만족도, 내용을 작성해주세요</S.StepDetail>
           </S.TopOff>
         )}
-
-        {/* STEP2 */}
         {props.activeStep === "second" && (
           <S.TopOn>
             <S.TopRow>
@@ -49,12 +41,10 @@ export default function WritePresenter(props: IWritePresenter) {
         )}
         {props.activeStep !== "second" && (
           <S.TopOff>
-            <S.StepName>STEP 2</S.StepName>
+            <S.StepName onClick={props.onClickFirstNext}>STEP 2</S.StepName>
             <S.StepDetail>가게가 어디에 있나요?</S.StepDetail>
           </S.TopOff>
         )}
-
-        {/* STEP3 */}
         {props.activeStep === "third" && (
           <S.TopOn>
             <S.TopRow>
@@ -66,270 +56,58 @@ export default function WritePresenter(props: IWritePresenter) {
         )}
         {props.activeStep !== "third" && (
           <S.TopOff>
-            <S.StepName>STEP 3</S.StepName>
+            <S.StepName onClick={props.onClickSecondNext}>STEP 3</S.StepName>
             <S.StepDetail>사진을 공유해주세요</S.StepDetail>
           </S.TopOff>
         )}
       </S.TopWrapper>
+      <First
+        activeStep={props.activeStep}
+        isEdit={props.isEdit}
+        data={props.data}
+        onChangeRange={props.onChangeRange}
+        startDate={props.startDate}
+        endDate={props.endDate}
+        dateError={props.dateError}
+        onChangeTitle={props.onChangeTitle}
+        title={props.title}
+        titleError={props.titleError}
+        resetScore={props.resetScore}
+        score={props.score}
+        onClickHappy={props.onClickHappy}
+        onClickUhm={props.onClickUhm}
+        onClickSad={props.onClickSad}
+        onClickResetScore={props.onClickResetScore}
+        scoreError={props.scoreError}
+        onChangeContents={props.onChangeContents}
+        contents={props.contents}
+        contentsError={props.contentsError}
+        isButtonActive={props.isButtonActive}
+        onClickFirstNext={props.onClickFirstNext}
+      />
+      <Second
+        activeStep={props.activeStep}
+        mapError={props.mapError}
+        mapStatus={props.mapStatus}
+        isEdit={props.isEdit}
+        data={props.data}
+        address={props.address}
+        isButtonActive={props.isButtonActive}
+        onClickSecondPrev={props.onClickSecondPrev}
+        onClickSecondNext={props.onClickSecondNext}
+      />
 
-      {/* STEP1 입력부분 */}
-      {props.activeStep === "first" && (
-        <S.BodyWrapper>
-          <S.Body>
-            <S.InputWrapper>
-              <S.Head>이 동네 거주 기간</S.Head>
-              <Space direction="vertical" size={12}>
-                {props.isEdit &&
-                  props.data?.fetchBoard.startDate &&
-                  props.data?.fetchBoard.endDate && (
-                    <S.StyledRangePicker
-                      onChange={props.onChangeRange}
-                      defaultValue={[
-                        moment(props.data?.fetchBoard.startDate, "YYYY-MM-DD"),
-                        moment(props.data?.fetchBoard.endDate, "YYYY-MM-DD"),
-                      ]}
-                    />
-                  )}
-                {!props.isEdit &&
-                  props.startDate !== "" &&
-                  props.endDate !== "" && (
-                    <S.StyledRangePicker
-                      onChange={props.onChangeRange}
-                      defaultValue={[
-                        moment(props.startDate, "YYYY-MM-DD"),
-                        moment(props.endDate, "YYYY-MM-DD"),
-                      ]}
-                    />
-                  )}
-                {!props.isEdit &&
-                  props.startDate === "" &&
-                  props.endDate === "" && (
-                    <S.StyledRangePicker onChange={props.onChangeRange} />
-                  )}
-              </Space>
-              <S.Error>{props.dateError}</S.Error>
-            </S.InputWrapper>
-            <S.InputWrapper>
-              <S.Head>글 제목</S.Head>
-              {props.isEdit ? (
-                <S.Input
-                  type="text"
-                  placeholder="제목을 입력해주세요."
-                  onChange={props.onChangeTitle}
-                  defaultValue={props.data?.fetchBoard.title || ""}
-                />
-              ) : (
-                <S.Input
-                  type="text"
-                  placeholder="제목을 입력해주세요."
-                  onChange={props.onChangeTitle}
-                  defaultValue={props.title || ""}
-                />
-              )}
-              <S.Error>{props.titleError}</S.Error>
-            </S.InputWrapper>
-            <S.InputWrapper>
-              <S.Head>만족도</S.Head>
-              {props.resetScore ? (
-                <S.RatingWrapper>
-                  {props.score === 1 ? (
-                    <S.RatingImg
-                      src="/image/happypick.png"
-                      onClick={props.onClickHappy}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/happy.png"
-                      onClick={props.onClickHappy}
-                    />
-                  )}
-                  {props.score === 2 ? (
-                    <S.RatingImg
-                      src="/image/uhmpick.png"
-                      onClick={props.onClickUhm}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/uhm.png"
-                      onClick={props.onClickUhm}
-                    />
-                  )}
-                  {props.score === 3 ? (
-                    <S.RatingImg
-                      src="/image/sadpick.png"
-                      onClick={props.onClickSad}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/sad.png"
-                      onClick={props.onClickSad}
-                    />
-                  )}
-                </S.RatingWrapper>
-              ) : (
-                <S.RatingWrapper>
-                  {props.score === 1 || props.data?.fetchBoard.score === 1 ? (
-                    <S.RatingImg
-                      src="/image/happypick.png"
-                      onClick={props.onClickHappy}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/happy.png"
-                      onClick={props.onClickHappy}
-                    />
-                  )}
-                  {props.score === 2 || props.data?.fetchBoard.score === 2 ? (
-                    <S.RatingImg
-                      src="/image/uhmpick.png"
-                      onClick={props.onClickUhm}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/uhm.png"
-                      onClick={props.onClickUhm}
-                    />
-                  )}
-                  {props.score === 3 || props.data?.fetchBoard.score === 3 ? (
-                    <S.RatingImg
-                      src="/image/sadpick.png"
-                      onClick={props.onClickSad}
-                    />
-                  ) : (
-                    <S.RatingImg
-                      src="/image/sad.png"
-                      onClick={props.onClickSad}
-                    />
-                  )}
-                </S.RatingWrapper>
-              )}
-              {props.isEdit && (
-                <S.ResetScore onClick={props.onClickResetScore}>
-                  초기화
-                </S.ResetScore>
-              )}
-              <S.Error>{props.scoreError}</S.Error>
-            </S.InputWrapper>
-            <S.InputWrapper>
-              <S.HeadContents>내용</S.HeadContents>
-              {props.isEdit && props.data?.fetchBoard.contents && (
-                <S.StyledQuill
-                  onChange={props.onChangeContents}
-                  formats={formats}
-                  modules={modules}
-                  defaultValue={props.data?.fetchBoard.contents}
-                />
-              )}
-              {!props.isEdit && (
-                <S.StyledQuill
-                  onChange={props.onChangeContents}
-                  formats={formats}
-                  modules={modules}
-                  defaultValue={props.contents || ""}
-                />
-              )}
-              <S.Error>{props.contentsError}</S.Error>
-            </S.InputWrapper>
-          </S.Body>
-          <S.ButtonWrapper>
-            <S.Button isButtonActive={props.isButtonActive}>취소</S.Button>
-            <S.Button
-              isButtonActive={props.isButtonActive}
-              onClick={props.onClickFirstNext}
-            >
-              &gt;
-            </S.Button>
-          </S.ButtonWrapper>
-        </S.BodyWrapper>
-      )}
-
-      {/* STEP2 입력부분*/}
-      {props.activeStep === "second" && (
-        <S.BodyWrapper>
-          <S.Body>
-            <S.MapWrapper>
-              {/* <S.MapHead>위치</S.MapHead> */}
-              <S.Error>{props.mapError}</S.Error>
-              <KeyWord mapStatus={props.mapStatus} />
-            </S.MapWrapper>
-            {props.address.group_name !== "" && (
-              <>
-                <S.MapHead>카테고리</S.MapHead>
-                {props.isEdit ? (
-                  <S.MapDetail>
-                    {props.data?.fetchBoard.category || ""}
-                  </S.MapDetail>
-                ) : (
-                  <S.MapDetail>{props.address.group_name}</S.MapDetail>
-                )}
-              </>
-            )}
-            <S.MapHead>상호명</S.MapHead>
-            {props.isEdit ? (
-              <S.MapDetail>{props.data?.fetchBoard.place || ""}</S.MapDetail>
-            ) : (
-              <S.MapDetail>{props.address.content}</S.MapDetail>
-            )}
-            <S.MapHead>주소</S.MapHead>
-            {props.isEdit ? (
-              <S.MapDetail>{props.data?.fetchBoard.address || ""}</S.MapDetail>
-            ) : (
-              <S.MapDetail>{props.address.address_name}</S.MapDetail>
-            )}
-          </S.Body>
-          <S.ButtonWrapper>
-            <S.Button
-              isButtonActive={props.isButtonActive}
-              onClick={props.onClickSecondPrev}
-            >
-              &lt;
-            </S.Button>
-            {props.address.address_name !== "" ? (
-              <S.MapButtonOn onClick={props.onClickSecondNext}>
-                &gt;
-              </S.MapButtonOn>
-            ) : (
-              <S.MapButtonOff onClick={props.onClickSecondNext}>
-                &gt;
-              </S.MapButtonOff>
-            )}
-          </S.ButtonWrapper>
-        </S.BodyWrapper>
-      )}
-
-      {/* STEP3 입력부분*/}
-      {props.activeStep === "third" && (
-        <S.ImageBodyWrapper>
-          <S.Body>
-            <S.ImageBody>
-              {props.fileUrls.map((el, index) => (
-                <ImageBoardUpload
-                  key={uuidv4()}
-                  index={index}
-                  fileUrl={el}
-                  onChangeFileUrls={props.onChangeFileUrls}
-                  // defaultValue={props.data?.fetchBoard.images?.imageUrl || ""}
-                />
-              ))}
-            </S.ImageBody>
-          </S.Body>
-          <S.ImageButtonWrapper>
-            <S.Button
-              isButtonActive={props.isButtonActive}
-              onClick={props.onClickThirdPrev}
-            >
-              &lt;
-            </S.Button>
-            <S.FinishButton
-              onClick={
-                props.isEdit ? props.onClickEditBoard : props.onClickWriteBoard
-              }
-            >
-              {props.isEdit ? "수정" : "등록"}
-            </S.FinishButton>
-          </S.ImageButtonWrapper>
-        </S.ImageBodyWrapper>
-      )}
+      <Third
+        activeStep={props.activeStep}
+        onChangeFileUrls={props.onChangeFileUrls}
+        fileUrls={props.fileUrls}
+        isButtonActive={props.isButtonActive}
+        onClickThirdPrev={props.onClickThirdPrev}
+        isEdit={props.isEdit}
+        onClickEditBoard={props.onClickEditBoard}
+        onClickWriteBoard={props.onClickWriteBoard}
+        data={props.data}
+      />
     </S.WrapperOut>
   );
 }
